pyproject.toml
[project]
name = "exchange-house"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
  "aerich==0.8.2",
  "fastapi[standard]==0.115.12",
  "pydantic-settings==2.8.1",
  "tortoise-orm[asyncpg]==0.24.2",
]

[tool.uv]
dev-dependencies = [
  "pytest==8.3.5",
  "pytest-asyncio==0.26.0",
  "mypy==1.15.0",
  "ruff==0.11.5",
  "coverage==7.8.0",
  "pre-commit==4.2.0",
  "tomlkit>=0.13.2",
  "pytest-env>=1.1.5",
]

[tool.mypy]
strict = true
exclude = [".venv", "alembic"]
python_version = "3.13"

[tool.ruff]
line-length = 120
exclude = ["alembic"]
target-version = "py313"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "W191"]

[tool.aerich]
tortoise_orm = "app.core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
